About
The program RandomWalker that takes an int command-line argument n
and simulates the motion of a random walk for n steps. Print the
location at each step (including the starting point), treating the
starting point as the origin (0, 0). Also, print the square of the
final Euclidean distance from the origin.